GSYE-678: Rename strategy profile related module files
