+++ src/gsy_e/models/strategy/energy_parameters/heat_pump.py
@@ -7,7 +7,7 @@ from gsy_framework.read_user_profile import InputProfileTypes
 from pendulum import DateTime
 
 from gsy_e.constants import FLOATING_POINT_TOLERANCE
-from gsy_e.models.strategy.profile import profile_factory
+from gsy_e.models.strategy.strategy_profile import profile_factory
 from gsy_e.models.strategy.state import HeatPumpState
 
 # pylint: disable=pointless-string-statement
