+++ src/gsy_e/models/strategy/strategy_profile.py
